+++ src/math/transformations.cpp
@@ -1,6 +1,6 @@
 #include "transformations.hpp"
 
-Transform Transform::translation(float x, float y, float z) {
+Transform Transform::translation(double x, double y, double z) {
     Transform translation;
     translation(0, 3) = x;
     translation(1, 3) = y;
@@ -8,7 +8,7 @@ Transform Transform::translation(float x, float y, float z) {
     return translation;
 }
 
-Transform Transform::scaling(float x, float y, float z) {
+Transform Transform::scaling(double x, double y, double z) {
     Transform scaling;
     scaling(0, 0) = x;
     scaling(1, 1) = y;
@@ -16,7 +16,7 @@ Transform Transform::scaling(float x, float y, float z) {
     return scaling;
 }
 
-Transform Transform::rotation_x(float rad) {
+Transform Transform::rotation_x(double rad) {
     Transform rotation_x;
     rotation_x(1, 1) = cos(rad);
     rotation_x(1, 2) = -sin(rad);
@@ -25,7 +25,7 @@ Transform Transform::rotation_x(float rad) {
     return rotation_x;
 }
 
-Transform Transform::rotation_y(float rad) {
+Transform Transform::rotation_y(double rad) {
     Transform rotation_y;
     rotation_y(0, 0) = cos(rad);
     rotation_y(0, 2) = sin(rad);
@@ -34,7 +34,7 @@ Transform Transform::rotation_y(float rad) {
     return rotation_y;
 }
 
-Transform Transform::rotation_z(float rad) {
+Transform Transform::rotation_z(double rad) {
     Transform rotation_z;
     rotation_z(0, 0) = cos(rad);
     rotation_z(0, 1) = -sin(rad);
@@ -43,8 +43,8 @@ Transform Transform::rotation_z(float rad) {
     return rotation_z;
 }
 
-Transform Transform::shearing(float x_y, float x_z, float y_x, float y_z,
-                                     float z_x, float z_y) {
+Transform Transform::shearing(double x_y, double x_z, double y_x, double y_z,
+    double z_x, double z_y) {
     Transform shearing;
     shearing(0, 1) = x_y;
     shearing(0, 2) = x_z;
@@ -61,9 +61,9 @@ Transform Transform::view_transform(Point from, Point to, Vector up) {
     Vector true_up = left.cross(forward);
 
     Matrix orientation(4, 4);
-    orientation = {left.x,     left.y,     left.z,     0,
+    orientation = { left.x,     left.y,     left.z,     0,
                    true_up.x,  true_up.y,  true_up.z,  0,
                    -forward.x, -forward.y, -forward.z, 0,
-                   0,          0,          0,          1};
+                   0,          0,          0,          1 };
     return Transform(orientation * translation(-from.x, -from.y, -from.z));
 }
